Update AWS dependencies, use default aws auth when AWS_* env is not present (#6)
We're seeing an issue with using this plugin in conjunction with AWS Creds backed by Nomad Workload Identities.

The issue is because we are forcing ec2 auth for AWS when really we should just trust the default auth provided by the SDK which tries a few methods for you. This fixes that error by just removing the explicit provider and using the implicit provider when some override variables are unset.

While I'm here, upgrading AWS dependencies for good measure. Tracking the deps used in the upstream nomad-autoscaler for dependency versions